+++ cellrank/tools/kernels/__init__.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
-from cellrank.tools.kernels._kernel import (
-    Kernel,
-    Constant,
-    PalantirKernel,
-    VelocityKernel,
-    PrecomputedKernel,
-    ConnectivityKernel,
-)
+from cellrank.tools.kernels._base_kernel import Kernel, Constant
+from cellrank.tools.kernels._palantir_kernel import PalantirKernel
+from cellrank.tools.kernels._velocity_kernel import VelocityKernel
+from cellrank.tools.kernels._precomputed_kernel import PrecomputedKernel
+from cellrank.tools.kernels._connectivity_kernel import ConnectivityKernel
