+++ cellrank/tools/kernels/_base_kernel.py
@@ -15,25 +15,20 @@ from typing import (
     Iterable,
     Optional,
 )
-from functools import wraps, reduce, partial
-
-from scvelo.preprocessing.moments import get_moments
+from functools import wraps, reduce
 
 import numpy as np
 from cellrank import logging as logg
 from scipy.sparse import spdiags, issparse, spmatrix, csr_matrix, isspmatrix_csr
 from cellrank.utils._docs import d
 from cellrank.tools._utils import (
-    bias_knn,
     _normalize,
     _get_neighs,
     _has_neighs,
-    _vals_to_csr,
     is_connected,
     is_symmetric,
-    _predict_transition_probabilities,
 )
-from cellrank.utils._utils import _read_graph_data, _write_graph_data
+from cellrank.utils._utils import _write_graph_data
 from cellrank.tools._constants import Direction, _transition
 
 _ERROR_DIRECTION_MSG = "Can only combine kernels that have the same direction."
@@ -215,7 +210,7 @@ class KernelExpression(ABC):
 
     def __rxor__(self, other: "KernelExpression") -> "KernelExpression":
         def convert(obj):
-            if isinstance(obj, _adaptive_add_type):
+            if _is_adaptive_type(obj):
                 if obj._mat_scaler is None:
                     raise ValueError(_ERROR_VAR_NOT_FOUND.format(obj))
                 return KernelMul(
@@ -226,7 +221,7 @@ class KernelExpression(ABC):
                         obj,
                     ]
                 )
-            if isinstance(_is_bin_mult(obj, return_constant=False), _adaptive_add_type):
+            if _is_adaptive_type(_is_bin_mult(obj, return_constant=False)):
                 e, c = _get_expr_and_constant(obj)
                 if e._mat_scaler is None:
                     raise ValueError(_ERROR_VAR_NOT_FOUND.format(e))
@@ -237,19 +232,15 @@ class KernelExpression(ABC):
             return obj
 
         if (
-            not isinstance(self, _adaptive_add_type)
+            not _is_adaptive_type(self)
             and not isinstance(self, KernelAdaptiveAdd)
-            and not isinstance(
-                _is_bin_mult(self, return_constant=False), _adaptive_add_type
-            )
+            and not _is_adaptive_type(_is_bin_mult(self, return_constant=False))
         ):
             raise TypeError(_ERROR_CONF_ADAPT.format(self.__class__.__name__))
         if (
-            not isinstance(other, _adaptive_add_type)
+            not _is_adaptive_type(other)
             and not isinstance(other, KernelAdaptiveAdd)
-            and not isinstance(
-                _is_bin_mult(other, return_constant=False), _adaptive_add_type
-            )
+            and not _is_adaptive_type(_is_bin_mult(other, return_constant=False))
         ):
             raise TypeError(_ERROR_CONF_ADAPT.format(other.__class__.__name__))
 
@@ -314,6 +305,7 @@ class KernelExpression(ABC):
     def __mul__(
         self, other: Union[float, int, "KernelExpression"]
     ) -> "KernelExpression":
+
         if isinstance(other, (int, float)):
             other = Constant(self.adata, other, backward=self.backward)
 
@@ -705,665 +697,6 @@ class ConstantMatrix(Kernel):
         return str(round(self._value, _n_dec))
 
 
-@d.dedent
-class PrecomputedKernel(Kernel):
-    """
-    Kernel which contains precomputed transition matrix.
-
-    Parameters
-    ----------
-    transition_matrix
-        Row-normalized transition matrix or a key in :paramref:`adata` `.obsp`.
-    %(adata)s
-    %(backward)s
-    """
-
-    def __init__(
-        self,
-        transition_matrix: Union[np.ndarray, spmatrix, str],
-        adata: Optional[AnnData] = None,
-        backward: bool = False,
-        compute_cond_num: bool = False,
-    ):
-        from anndata import AnnData as _AnnData
-
-        if isinstance(transition_matrix, str):
-            if adata is None:
-                raise ValueError(
-                    "When `transition_matrix` specifies a key to `adata.obsp`, `adata` cannot be None."
-                )
-            transition_matrix = _read_graph_data(adata, transition_matrix)
-
-        if not isinstance(transition_matrix, (np.ndarray, spmatrix)):
-            raise TypeError(
-                f"Expected transition matrix to be of type `numpy.ndarray` or `scipy.sparse.spmatrix`, "
-                f"found `{type(transition_matrix).__name__!r}`."
-            )
-
-        if transition_matrix.shape[0] != transition_matrix.shape[1]:
-            raise ValueError(
-                f"Expected transition matrix to be square, found `{transition_matrix.shape}`."
-            )
-
-        if not np.allclose(np.sum(transition_matrix, axis=1), 1.0):
-            raise ValueError("Not a valid transition matrix: not all rows sum to 1.")
-
-        if adata is None:
-            logg.debug("Creating empty dummy AnnData object")
-            adata = _AnnData(
-                csr_matrix((transition_matrix.shape[0], 1), dtype=np.float32)
-            )
-
-        super().__init__(adata, backward=backward, compute_cond_num=compute_cond_num)
-        self._transition_matrix = csr_matrix(transition_matrix)
-        self._maybe_compute_cond_num()
-
-    def _read_from_adata(self, **kwargs):
-        pass
-
-    def copy(self) -> "PrecomputedKernel":
-        """Return a copy of self."""
-        pk = PrecomputedKernel(
-            copy(self.transition_matrix), adata=self.adata, backward=self.backward
-        )
-        pk._cond_num = self.condition_number
-        return pk
-
-    def compute_transition_matrix(self, *args, **kwargs) -> "PrecomputedKernel":
-        """Return self."""
-        return self
-
-    def __invert__(self) -> "PrecomputedKernel":
-        # do not call parent's invert, since it removes the transition matrix
-        self._direction = Direction.FORWARD if self.backward else Direction.BACKWARD
-        return self
-
-    def __repr__(self):
-        return f"{'~' if self.backward and self._parent is None else ''}<Precomputed>"
-
-    def __str__(self):
-        return repr(self)
-
-
-@d.dedent
-class VelocityKernel(Kernel):
-    """
-    Kernel which computes a transition matrix based on velocity correlations.
-
-    This borrows ideas from both [Manno18]_ and [Bergen19]_. In short, for each cell *i*, we compute transition
-    probabilities :math:`p_{i, j}` to each cell *j* in the neighborhood of *i*. The transition probabilities are
-    computed as a multinominal logistic regression where the weights :math:`w_j` (for all *j*) are given by the vector
-    that connects cell *i* with cell *j* in gene expression space, and the features :math:`x_i` are given
-    by the velocity vector :math:`v_i` of cell *i*.
-
-    Optionally, we propagate uncertainty in the velocity vectors forward into transition probabilities using either
-    an analytical approximation or a Monte Carlo approach.
-
-    Parameters
-    ----------
-    %(adata)s
-    %(backward)s
-    vkey
-        Key in :paramref:`adata` `.uns` where the velocities are stored.
-    xkey
-        Key in :paramref:`adata` `.layers` where expected gene expression counts are stored.
-    gene_subset
-        List of genes to be used to compute transition probabilities. By default, the `velocity_genes` of
-        :paramref:`adata` `. var` are used.
-    compute_cond_num
-        Whether to compute condition number of the transition matrix. Note that this might be costly,
-        since it does not use sparse implementation.
-    check_connectivity
-        Check whether the underlying KNN graph is connected.
-    """
-
-    def __init__(
-        self,
-        adata: AnnData,
-        backward: bool = False,
-        vkey: str = "velocity",
-        xkey: str = "Ms",
-        gene_subset: Optional[Iterable] = None,
-        compute_cond_num: bool = False,
-        check_connectivity: bool = False,
-    ):
-        super().__init__(
-            adata,
-            backward=backward,
-            vkey=vkey,
-            xkey=xkey,
-            gene_subset=gene_subset,
-            compute_cond_num=compute_cond_num,
-            check_connectivity=check_connectivity,
-        )
-        self._vkey = vkey  # for copy
-        self._xkey = xkey
-        self._gene_subset = gene_subset
-        self._pearson_correlations = None
-
-    def _read_from_adata(self, **kwargs):
-        super()._read_from_adata(**kwargs)
-
-        # check whether velocities have been computed
-        vkey = kwargs.pop("vkey", "velocity")
-        if vkey not in self.adata.layers.keys():
-            raise KeyError("Compute RNA velocity first as `scv.tl.velocity()`.")
-
-        # restrict genes to a subset, i.e. velocity genes or user provided list
-        gene_subset = kwargs.pop("gene_subset", None)
-        subset = np.ones(self.adata.n_vars, bool)
-        if gene_subset is not None:
-            var_names_subset = self.adata.var_names.isin(gene_subset)
-            subset &= var_names_subset if len(var_names_subset) > 0 else gene_subset
-        elif f"{vkey}_genes" in self.adata.var.keys():
-            subset &= np.array(self.adata.var[f"{vkey}_genes"].values, dtype=bool)
-
-        # chose data representation to use for transcriptomic displacements
-        xkey = kwargs.pop("xkey", "Ms")
-        xkey = xkey if xkey in self.adata.layers.keys() else "spliced"
-
-        # filter both the velocities and the gene expression profiles to the gene subset. Densify the matrices.
-        X = np.array(
-            self.adata.layers[xkey].A[:, subset]
-            if issparse(self.adata.layers[xkey])
-            else self.adata.layers[xkey][:, subset]
-        )
-        V = np.array(
-            self.adata.layers[vkey].A[:, subset]
-            if issparse(self.adata.layers[vkey])
-            else self.adata.layers[vkey][:, subset]
-        )
-
-        # remove genes that have any Nan values (in both X and V)
-        nans = np.isnan(np.sum(V, axis=0))
-        if np.any(nans):
-            X = X[:, ~nans]
-            V = V[:, ~nans]
-
-        # check the velocity parameters
-        if vkey + "_params" in self.adata.uns.keys():
-            velocity_params = self.adata.uns[vkey + "_params"]
-        else:
-            velocity_params = None
-            logg.debug("Unable to load velocity parameters")
-
-        # add to self
-        self._velocity = V
-        self._gene_expression = X
-        self._velocity_params = velocity_params
-
-    def compute_transition_matrix(
-        self,
-        backward_mode: str = "transpose",
-        softmax_scale: float = 4.0,
-        mode: str = "deterministic",
-        seed: Optional[int] = None,
-        **kwargs,
-    ) -> "VelocityKernel":
-        """
-        Compute transition matrix based on velocity directions on the local manifold.
-
-        For each cell, infer transition probabilities based on the correlation of the cell's
-        velocity-extrapolated cell state with cell states of its *K* nearest neighbors.
-
-        Parameters
-        ----------
-        backward_mode
-            Options are `['transpose', 'negate']`. Only matters if initialized as :paramref:`backward` =`True`.
-        softmax_scale
-            Scaling parameter for the softmax.
-        mode
-            How to compute transition probabilities. Options are "stochastic" (propagate uncertainty analytically),
-            "deterministic" (don't propagate uncertainty) and "sampling" (sample from velocity distribution).
-        seed
-            Set the seed for random state, only relevant for `mode='sampling'`.
-
-        Returns
-        -------
-        None
-            Nothing, just makes available the following fields:
-
-                - :paramref:`transition_matrix`
-                - :paramref:`pearson_correlations`
-        """
-
-        if mode not in ["stochastic", "deterministic", "sampling"]:
-            raise NotImplementedError(f"Mode `{mode}` is not implemented.")
-
-        if mode != "deterministic" and self._direction == Direction.BACKWARD:
-            logg.warning(
-                f"Mode `{mode}` is currently not supported for the backward process. "
-                f"Defaulting to mode `'deterministic'`"
-            )
-            mode = "deterministic"
-
-        start = logg.info("Computing transition matrix based on velocity correlations")
-
-        params = dict(  # noqa
-            bwd_mode=backward_mode if self._direction == Direction.BACKWARD else None,
-            sigma_corr=softmax_scale,
-            mode=mode,
-            random_state=seed,
-        )
-
-        # check whether we already computed such a transition matrix. If yes, load from cache
-        if params == self._params:
-            assert self.transition_matrix is not None, _ERROR_EMPTY_CACHE_MSG
-            logg.debug(_LOG_USING_CACHE, time=start)
-            logg.info("    Finish", time=start)
-            return self
-
-        self._params = params
-
-        # compute first and second order moments to model the distribution of the velocity vector
-        if mode in ["stochastic", "sampling"]:
-            np.random.seed(seed)
-            velocity_expectation = get_moments(
-                self.adata, self._velocity, second_order=False
-            )
-            velocity_variance = get_moments(
-                self.adata, self._velocity, second_order=True
-            )
-
-        # define a function to compute hessian matrices
-        if mode == "stochastic":
-            try:
-                from jax import jit, jacfwd, jacrev
-
-                get_hessian_fwd = jit(
-                    jacfwd(
-                        jacrev(
-                            partial(
-                                _predict_transition_probabilities,
-                                return_pearson_correlation=False,
-                                use_jax=True,
-                            )
-                        )
-                    )
-                )
-            except ImportError:
-                logg.warning("TODO: install jax")
-                mode = "sampling"
-
-        # sort cells by their number of neighbors - this makes jitting more efficient
-        n_neighbors = np.array((self._conn != 0).sum(1)).flatten()
-        cell_ixs = np.argsort(n_neighbors)[::-1]
-
-        # loop over all cells
-        probs_list, corrs_list, rows, cols, n_obs = (
-            [],
-            [],
-            [],
-            [],
-            self._gene_expression.shape[0],
-        )
-        for cell_ix in cell_ixs:
-
-            # get the neighbors
-            nbhs_ixs = self._conn[cell_ix, :].indices
-
-            # get the displacement matrix. Changing dimensions b/c varying numbers of neighbors slow down autograd
-            W = self._gene_expression[nbhs_ixs, :] - self._gene_expression[cell_ix, :]
-
-            if mode == "deterministic":
-                # treat `v_i` as deterministic, with no error
-
-                if self._direction == Direction.FORWARD or (
-                    self._direction == Direction.BACKWARD and backward_mode == "negate"
-                ):
-                    # evaluate the prediction at the actual velocity vector
-                    v_i = self._velocity[cell_ix, :]
-
-                    # for the transpose backward mode, just flip the velocity vector
-                    if self._direction == Direction.BACKWARD:
-                        v_i *= -1
-
-                    if (v_i == 0).all():
-                        logg.warning(f"Cell `{cell_ix}` has a zero velocity vector")
-                        probs, corrs = (
-                            np.ones(len(nbhs_ixs)) / len(nbhs_ixs),
-                            np.zeros(len(nbhs_ixs)),
-                        )
-                    else:
-                        probs, corrs = _predict_transition_probabilities(
-                            v_i, W, softmax_scale=softmax_scale,
-                        )
-                else:
-                    # compute how likely all neighbors are to transition to this cell
-                    V = self._velocity[nbhs_ixs, :]
-                    probs, corrs = _predict_transition_probabilities(
-                        V, -1 * W, softmax_scale=softmax_scale,
-                    )
-
-            elif mode == "stochastic":
-                # treat `v_i` as random variable and use analytical approximation to propagate the distribution
-
-                # get the expected velocity vector
-                v_i = velocity_expectation[cell_ix, :]
-
-                if (v_i == 0).all():
-                    logg.warning(f"Cell `{cell_ix}` has a zero velocity vector")
-                    probs, corrs = (
-                        np.ones(len(nbhs_ixs)) / len(nbhs_ixs),
-                        np.zeros(len(nbhs_ixs)),
-                    )
-                else:
-                    # get the variance of the distribution over velocity vectors
-                    variances = velocity_variance[cell_ix, :]
-
-                    # compute the Hessian tensor, and turn it into a matrix that has the diagonal elements in its rows
-                    H = get_hessian_fwd(v_i, W, softmax_scale)
-                    H_diag = np.concatenate([np.diag(h)[None, :] for h in H])
-
-                    # compute zero order term
-                    p_0, corrs = _predict_transition_probabilities(
-                        v_i, W, softmax_scale=softmax_scale,
-                    )
-
-                    # compute second order term (note that the first order term cancels)
-                    p_2 = 0.5 * H_diag.dot(variances)
-
-                    # combine both to give the second order Taylor approximation. Can sometimes be negative because we
-                    # neglected higher order terms, so force it to be non-negative
-                    probs = np.where((p_0 + p_2) >= 0, p_0 + p_2, 0)
-
-            elif mode == "sampling":
-                # treat `v_i` as random variable and use Monte Carlo approximation to propagate the distribution
-
-                # sample a single velocity vector from the distribution
-                v_i = np.random.normal(
-                    velocity_expectation[cell_ix, :], velocity_variance[cell_ix, :]
-                )
-
-                # use this sample to compute transition probabilities
-                probs, corrs = _predict_transition_probabilities(
-                    v_i, W, softmax_scale=softmax_scale,
-                )
-
-            # add the computed transition probabilities and correlations to lists
-            corrs_list.extend(corrs)
-            probs_list.extend(probs)
-            rows.extend(np.ones(len(nbhs_ixs)) * cell_ix)
-            cols.extend(nbhs_ixs)
-
-        # collect everything in a matrix of correlations and one of transition probabilities
-        probs_matrix, corrs_matrix = np.hstack(probs_list), np.hstack(corrs_list)
-        probs_matrix[np.isnan(probs_matrix)] = 0
-        corrs_matrix[np.isnan(corrs_matrix)] = 0
-
-        # transform to sparse matrices
-        corrs_matrix = _vals_to_csr(corrs_matrix, rows, cols, shape=(n_obs, n_obs))
-        probs_matrix = _vals_to_csr(probs_matrix, rows, cols, shape=(n_obs, n_obs))
-
-        self._compute_transition_matrix(probs_matrix, density_normalize=False)
-        self._pearson_correlations = corrs_matrix
-
-        logg.info("    Finish", time=start)
-
-        return self
-
-    @property
-    def pearson_correlations(self) -> csr_matrix:
-        """TODO."""
-        return self._pearson_correlations
-
-    def copy(self) -> "VelocityKernel":
-        """Return a copy of self."""
-        vk = VelocityKernel(
-            self.adata,
-            backward=self.backward,
-            vkey=self._vkey,
-            xkey=self._xkey,
-            gene_subset=self._gene_subset,
-        )
-        vk._params = copy(self.params)
-        vk._cond_num = self.condition_number
-        vk._transition_matrix = copy(self._transition_matrix)
-        vk._pearson_correlations = copy(self.pearson_correlations)
-
-        return vk
-
-
-@d.dedent
-class ConnectivityKernel(Kernel):
-    """
-    Kernel which computes transition probabilities based on transcriptomic similarities.
-
-    As a measure for transcriptomic similarity, we use the weighted KNN graph computed using
-    :func:`scanpy.pp.neighbors`,see [Wolf18]_.
-    By definition, the resulting transition matrix is symmetric and cannot be used to learn about the direction of the
-    developmental process under consideration.
-    However, the velocity-derived transition matrix can be combined with the similarity-based transition matrix as
-    a means of regularization.
-
-    Optionally, we apply a density correction as described in [Coifman05]_, where we use the implementation of
-    [Haghverdi16]_.
-
-    Parameters
-    ----------
-    %(adata)s
-    %(backward)s
-    compute_cond_num
-        Whether to compute condition number of the transition matrix. Note that this might be costly,
-        since it does not use sparse implementation.
-    check_connectivity
-        Check whether the underlying KNN graph is connected.
-    """
-
-    def __init__(
-        self,
-        adata: AnnData,
-        backward: bool = False,
-        compute_cond_num: bool = False,
-        check_connectivity: bool = False,
-    ):
-        super().__init__(
-            adata,
-            backward=backward,
-            compute_cond_num=compute_cond_num,
-            check_connectivity=check_connectivity,
-        )
-
-    def compute_transition_matrix(
-        self, density_normalize: bool = True
-    ) -> "ConnectivityKernel":
-        """
-        Compute transition matrix based on transcriptomic similarity.
-
-        Uses symmetric, weighted KNN graph to compute symmetric transition matrix. The connectivities are computed
-        using :func:`scanpy.pp.neighbors`. Depending on the parameters used there, they can be UMAP connectivities or
-        gaussian-kernel-based connectivities with adaptive kernel width.
-
-        Parameters
-        ----------
-        density_normalize
-            Whether or not to use the underlying KNN graph for density normalization.
-
-        Returns
-        -------
-        None
-            Nothing, but makes :paramref:`transition_matrix` available.
-        """
-
-        start = logg.info("Computing transition matrix based on connectivities")
-
-        params = {"dnorm": density_normalize}
-        if params == self.params:
-            assert self.transition_matrix is not None, _ERROR_EMPTY_CACHE_MSG
-            logg.debug(_LOG_USING_CACHE)
-            logg.info("    Finish", time=start)
-            return self
-
-        self._params = params
-        self._compute_transition_matrix(
-            matrix=self._conn.copy(), density_normalize=density_normalize
-        )
-
-        logg.info("    Finish", time=start)
-
-        return self
-
-    def copy(self) -> "ConnectivityKernel":
-        """Return a copy of self."""
-        ck = ConnectivityKernel(self.adata, backward=self.backward,)
-        ck._params = copy(self.params)
-        ck._cond_num = self.condition_number
-        ck._transition_matrix = copy(self._transition_matrix)
-
-        return ck
-
-
-@d.dedent
-class PalantirKernel(Kernel):
-    """
-    Kernel which computes transition probabilities in a similar way to *Palantir*, see [Setty19]_.
-
-    *Palantir* computes a KNN graph in gene expression space and a pseudotime, which it then uses to direct the edges of
-    the KNN graph, such that they are more likely to point into the direction of increasing pseudotime. To avoid
-    disconnecting the graph, it does not remove all edges that point into the direction of decreasing pseudotime
-    but keeps the ones that point to nodes inside a close radius. This radius is chosen according to the local density.
-
-    The implementation presented here won't exactly reproduce the original *Palantir* algorithm (see below)
-    but the results are qualitatively very similar.
-
-    Optionally, we apply density correction as described in [Coifman05]_, where we use the implementation of
-    [Haghverdi16]_.
-
-    Parameters
-    ----------
-    %(adata)s
-    %(backward)s
-    time_key
-        Key in :paramref:`adata` `.obs` where the pseudotime is stored.
-    compute_cond_num
-        Whether to compute condition number of the transition matrix. Note that this might be costly,
-        since it does not use sparse implementation.
-    """
-
-    def __init__(
-        self,
-        adata: AnnData,
-        backward: bool = False,
-        time_key: str = "dpt_pseudotime",
-        compute_cond_num: bool = False,
-        check_connectivity: bool = False,
-    ):
-        super().__init__(
-            adata,
-            backward=backward,
-            time_key=time_key,
-            compute_cond_num=compute_cond_num,
-            check_connectivity=check_connectivity,
-        )
-        self._time_key = time_key
-
-    def _read_from_adata(self, **kwargs):
-        super()._read_from_adata(**kwargs)
-
-        time_key = kwargs.pop("time_key", "dpt_pseudotime")
-        if time_key not in self.adata.obs.keys():
-            raise KeyError(f"Could not find time key in `adata.obs[{time_key!r}]`.")
-        logg.debug("Adding `.pseudotime`")
-
-        self._pseudotime = np.array(self.adata.obs[time_key]).astype(_dtype)
-
-        if np.nanmin(self.pseudotime) < 0:
-            raise ValueError(
-                f"Minimum pseudotime must be non-negative, found {np.nanmin(self.pseudotime)}."
-            )
-
-    def compute_transition_matrix(
-        self, k: int = 3, density_normalize: bool = True
-    ) -> "PalantirKernel":
-        """
-        Compute transition matrix based on KNN graph and pseudotemporal ordering.
-
-        This is a re-implementation of the Palantir algorithm by [Setty19]_.
-        Note that this won't exactly reproduce the original Palantir results, for three reasons:
-
-            - 1. Palantir computes the KNN graph in a scaled space of diffusion components
-            - 2. Palantir uses its own pseudotime to bias the KNN graph which is not implemented here
-            - 3. Palantir uses a slightly different mechanism to ensure the graph remains connected when removing edges that point into the "pseudotime past"
-
-        If you would like to reproduce the original results, please use the original Palantir algorithm.
-
-        Parameters
-        ----------
-        k
-            :paramref:`k` is the number of neighbors to keep for each node, regardless of pseudotime.
-            This is done to ensure that the graph remains connected.
-        density_normalize
-            Whether or not to use the underlying KNN graph for density normalization.
-
-        Returns
-        -------
-        None
-            Nothing, but makes :paramref:`transition_matrix` available.
-        """  # noqa
-
-        start = logg.info("Computing transition matrix based on Palantir-like kernel")
-
-        # get the connectivities and number of neighbors
-        if (
-            "neighbors" in self.adata.uns
-            and "params" in self.adata.uns["neighbors"]
-            and "n_neighbors" in self.adata.uns["neighbors"]["params"]
-        ):
-            n_neighbors = self.adata.uns["neighbors"]["params"]["n_neighbors"]
-        else:
-            logg.warning(
-                "Could not find 'n_neighbors' in `adata.uns['neighbors']['params']`. Using an estimate"
-            )
-            n_neighbors = np.min(self._conn.sum(1))
-
-        params = dict(k=k, dnorm=density_normalize, n_neighs=n_neighbors)  # noqa
-        if params == self._params:
-            assert self.transition_matrix is not None, _ERROR_EMPTY_CACHE_MSG
-            logg.debug(_LOG_USING_CACHE)
-            logg.info("    Finish", time=start)
-            return self
-
-        self._params = params
-
-        # handle backward case and run biasing function
-        pseudotime = (
-            np.max(self.pseudotime) - self.pseudotime
-            if self._direction == Direction.BACKWARD
-            else self.pseudotime
-        )
-        biased_conn = bias_knn(
-            conn=self._conn, pseudotime=pseudotime, n_neighbors=n_neighbors, k=k
-        ).astype(_dtype)
-        # make sure the biased graph is still connected
-        if not is_connected(biased_conn):
-            logg.warning("Biased KNN graph is disconnected")
-
-        self._compute_transition_matrix(
-            matrix=biased_conn, density_normalize=density_normalize
-        )
-
-        logg.info("    Finish", time=start)
-
-        return self
-
-    @property
-    def pseudotime(self) -> np.array:
-        """Pseudotemporal ordering of cells."""
-        return self._pseudotime
-
-    def copy(self) -> "PalantirKernel":
-        """Return a copy of self."""
-        pk = PalantirKernel(
-            self.adata, backward=self.backward, time_key=self._time_key,
-        )
-        pk._pseudotime = copy(self.pseudotime)
-        pk._params = copy(self._params)
-        pk._cond_num = self.condition_number
-        pk._transition_matrix = copy(self._transition_matrix)
-
-        return pk
-
-
 class SimpleNaryExpression(NaryKernelExpression):
     """Base class for n-ary operations."""
 
@@ -1540,9 +873,5 @@ def _is_bin_mult(
     return None
 
 
-_adaptive_add_type = (
-    ConnectivityKernel,
-    VelocityKernel,
-    PalantirKernel,
-    PrecomputedKernel,
-)
+def _is_adaptive_type(k: KernelExpression) -> bool:
+    return isinstance(k, Kernel) and not isinstance(k, ((Constant, ConstantMatrix)))
